DjangoRestAPI-project/backendAPI/requirements.txt
Django
Pillow
djangorestframework
django-cors-headers
python-decouple
mysqlclient